Working projection with 1/3 rotations
Right now im able to project a point from 3 dimensional space
onto the screen. Great success!

I can also rotate a shape around an axis. When rotating around
the Z axis it works as expected, but using the rotations around
the X and Y axis makes it so the shape gets rotated around the
viewer instead of the shapes' own axis. I do not know why. It
could be that the rotation matrices are not adjusted for
coordinates in object space.. but I don't know yet

I've been very happy with how I've been able to use the type
system to ensure the correct units are used everywhere. When
writing my 2D game in Python I would often run into stupid bugs
where I would confuse an angle for a radian. Since I now use
the type system to keep track of this, I can be much more sure
that my program is correct :)

Future work:

Change my method of projection to using a projection matrix. I'm
being told it's a better way to do it. While doing that I think
I'll be able to implement a 'camera' aswell, which will have its'
own position.

This is very future work, but it would be cool to be able to create
a shape from the application aswell. It should be possible to click
a point on the screen, and perform the projection "backwards" to
create a point in object space..?